+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - `npm start` not working anymore
+- Crash when using the Player Android SDK where `player.ads.isLinearAdActive` was invoked despite not being available on that platform
 
 ## [4.1.0] - 2025-10-06